Clipping-project/requirements.txt
opencv-python
numpy
psutil
albumentations
scikit-image
tqdm
moviepy
pydub
SpeechRecognition
librosa
pyAudioAnalysis
nltk
flask
ffmpeg-python